Create box_storage_accounts.py
This smart contract uses the Beaker framework to create a box called "accounts", which is a dictionary that stores the balances of user accounts. The box has two functions: "add_account" and "get_balance".

The "add_account" function takes in an account ID and a balance, and adds the account to the dictionary if it does not already exist. The "get_balance" function takes in an account ID and returns the balance of the account if it exists in the dictionary, or 0 if it does not.

In the main function, the "accounts" box is initialized and then used to add a few accounts with initial balances. The balances of these accounts are then printed to the console to demonstrate the functionality of the "get_balance" function.

By leveraging box storage, this smart contract is able to store and retrieve an unlimited number of accounts and their balances, which would not be possible with global or local storage due to their limited size and scope.